move data-viz components to new data module
