+++ src/once-ui/modules/data/BarGraph.module.scss
