+++ src/once-ui/modules/data/LineGraph.module.scss
